Simplify semantic assembly and declaration logic
- Flattened `src/semantic/assembly/` into `src/semantic/assembly.rs` for reduced indirection.
- Merged `ParticipleConstituent` into `Constituent`, removing redundant struct.
- Removed `Literal` enum, using `Expr` variants directly in `AssembledStatement` to eliminate type duplication.
- Refactored `extract_parameters_from_expr` to use a peekable iterator.
- Updated all call sites and tests to reflect these changes.
- Recorded simplifications in `.jules/razor.md`.
