+++ src/semantic/mod.rs
@@ -50,7 +50,7 @@ mod types;
 pub use crate::morphology::{DisambiguationContext, analyze_article, disambiguate, resolve_best};
 pub(crate) use assembly::Assembler;
 pub use assembly::{
-    AssembledStatement, AssemblyError, Constituent, Literal, ParticipleConstituent, VerbConstituent,
+    AssembledStatement, AssemblyError, Constituent, ParticipleConstituent, VerbConstituent,
 };
 pub use model::*;
 pub use resolver::*;
